# Improve `elastic-agent install` performance