--- a/go.mod
+++ b/go.mod
@@ -175,7 +175,7 @@ replace (
 	github.com/tonistiigi/fifo => github.com/containerd/fifo v0.0.0-20190816180239-bda0ff6ed73c
 )
 
-replace github.com/otiai10/copy => github.com/ycombinator/copy v0.0.0-20230831213531-be8664b0cabf
+replace github.com/otiai10/copy => github.com/ycombinator/copy v0.0.0-20230831215828-55a4d5bcc092
 
 // Exclude this version because the version has an invalid checksum.
 exclude github.com/docker/distribution v2.8.0+incompatible