--- a/internal/pkg/agent/install/install.go
+++ b/internal/pkg/agent/install/install.go
@@ -68,8 +68,8 @@ func Install(cfgFile, topPath string, pt ProgressTrackerStep) error {
 		OnSymlink: func(_ string) copy.SymlinkAction {
 			return copy.Shallow
 		},
-		Sync:        true,
-		Concurrency: int64(copyConcurrency),
+		Sync:         true,
+		NumOfWorkers: int64(copyConcurrency),
 	})
 	if err != nil {
 		s.Failed()
@@ -250,6 +250,7 @@ func verifyDirectory(dir string) error {
 func HasAllSSDs() bool {
 	block, err := ghw.Block()
 	if err != nil {
+		fmt.Fprintf(os.Stdout, "HasAllSSDs: ghw.Block() returned error: %s\n", err.Error())
 		return false
 	}
 
@@ -257,14 +258,18 @@ func HasAllSSDs() bool {
 		switch disk.DriveType {
 		case ghw.DRIVE_TYPE_FDD, ghw.DRIVE_TYPE_ODD:
 			// Floppy or optical drive; we don't care about these
+			fmt.Fprintf(os.Stdout, "HasAllSSDs: %s is a FDD or ODD\n", disk.Name)
 			continue
 		case ghw.DRIVE_TYPE_SSD:
 			// SSDs
+			fmt.Fprintf(os.Stdout, "HasAllSSDs: %s is a SSD\n", disk.Name)
 			continue
 		case ghw.DRIVE_TYPE_HDD:
 			// HDD (spinning hard disk)
+			fmt.Fprintf(os.Stdout, "HasAllSSDs: %s is a HDD\n", disk.Name)
 			return false
 		default:
+			fmt.Fprintf(os.Stdout, "HasAllSSDs: %s is of unknown type\n", disk.Name)
 			return false
 		}
 	}