--- a/internal/pkg/agent/install/install_test.go
+++ b/internal/pkg/agent/install/install_test.go
@@ -1,0 +1,1 @@
+package install