@ycombinator 2023-09-11
Currently waiting on https://github.com/otiai10/copy/pull/123 to be merged, then will update `go.mod` on this PR.