@ycombinator 2023-10-02
> Hey @michalpristas, I was wondering if you have a Windows machine with an HDD (as opposed to SSD). @pierrehilbert suggested that you might.
> 
> If you do, would you mind testing the performance of `elastic-agent install` on that machine before and after this PR? You can follow the same steps that I did for [Windows - SSD](https://github.com/elastic/elastic-agent/pull/3212#issue-1842215062) (scroll down a bit).
> 
> If you don't have such a machine, just let me know. Thanks either way!

Chatted with @michalpristas off-PR.  He doesn't have an HDD machine where `elastic-agent` can be installed.