@ycombinator 2023-10-05
@cmacknz Worth backporting to `8.11`?