@ycombinator 2023-10-31
> Missed the original ping here, not a bug fix so we shouldn't have backported.

👍 This PR was not backported so we're good.  Thanks!